refactor: improve shell components structure
Refactored how we did structure the shell components before. Now, we
have all the validation logic, which is the same, in the parent
component. We implement a variant of a factory method that allows us to
create one instance or the other depending on the option entered by the
user.
Also, we have a builder class that given a group of params will create
the common command configuration, which is all the same for the three
options (user, team and organization).
With this we have a cleaner code, with less repeated builder methods in
the entry point of the application.
Also, a description has been added to the dates so that the user knows
which date format is required for the `from` and `to` params.